Release v1.0 (bug fixes, configuration improvements)

+++ src/Models/PostData.h
@@ -18,5 +18,5 @@ struct PostData {
     uint8_t powerMode;
 
     unsigned long restartsCount;
-    unsigned int gsmErrors;
+    unsigned long gsmErrors;
 };
